Analyse a list of data blocks containing shape information

Performs either regularised consensus principal component analysis on a list of data blocks containing shape information, or performs principal component analysis on a superblock produced by the column-wise concatenation of the individual data blocks.

analyseBlocks(blockList, option = "rcpca", ncomp = 3)

Arguments

blockList
list of 'block' objects produced by the dodecBlock, formatBlock, readPts or readGPSA functions.

option
either option = "rcpca" (default) for regularised consensus principal component analysis in mode 2 applied to the entire block list, or option = "pca" for principal component analysis applied to the superblock item in the block list.

ncomp
an integer specifying how many components should be calculated (default is 3)

Value

A list object containing output from the regularised consensus principal component analysis or principal component analysis. The list contains the elements:

result

output from the regularised consensus principal component analysis in mode 2 produced by the rgcca function from the RGCCA package, or output from principal component analysis produced by the prcomp function in base R.

option

either "rcpca" or "pca".

block.list

a list containing the data blocks and a concatenated superblock. Inherited from the supplied blockList object and retained for downstream analyses.

scores

component score values (for individual blocks and the consensus if option = "rpca"; for the superblock if option = "pca") (see scoresPlot for more detail).

block.loadings

component loadings (for individual blocks and the consensus if option = "rpca"; for the superblock if option = "pca") (see loadingsPlot for more detail).

p

number of points in the configurations of each data block. Inherited from the supplied blockList object and retained for downstream analyses.

k

number of dimensions that the points in each configuration has. Inherited from the supplied blockList object and retained for downstream analyses.

n

number of configurations included in each data block. Inherited from the supplied blockList object and retained for downstream analyses.

Details

analyseBlocks is applied to an object of class "blockList" produced by the combineBlocks function and has two options: 1) option = "rcpca" and 2) option = "pca". The option = "rcpca" will perform regularised consensus principal component analysis using the rgcca function from the RGCCA package (Tenenhaus and Guillemot 2017), and is the default option for analyseBlocks. The rgcca function itself has many options that each perform a different type of analysis. Here the analyseBlocks function is specifically calling the regularised consensus principal component analysis in mode 2 option with scaling applied. For further detail see Tenenhaus and Guillemot (2017) and Tenenhaus et al. (2017). option = "pca" will perform principal component analysis on the superblock item in the block list using the prcomp function from base R.

References

Tenenhaus A, Guillemot V. 2017. RGCCA: Regularized and sparse generalized canonical correlation analysis for multiblock data 2.1.2. https://CRAN.R-project.org/package=RGCCA.

Tenenhaus M, Tenenhaus A, Groenen PJF. 2017. Regularized generalized canonical correlation analysis: A framework for sequential multiblock component methods. Psychometrika 82: 737-777 https://doi.org/10.1007/s11336-017-9573-x

Examples

block1 <- dodecBlock()
block2 <- dodecBlock()
blocklist <- combineBlocks(blocks = c(block1, block2))
result1 <- analyseBlocks(blocklist)
result2 <- analyseBlocks(blocklist, option = "pca", ncomp = 10)